+++ src/moving_avg_ticker.h
@@ -16,11 +16,11 @@ extern "C" {
 
 typedef struct moving_avg_values {
     // Circular buffer
-    int32_t *wbuf;
+    int16_t *wbuf;
     uint32_t wbuf_end;
     uint32_t wbuf_len;
 
-    int32_t subtotal;
+    int16_t subtotal;
     uint32_t subtotal_size;
     uint32_t max_subtotal_size;
 } moving_avg_values;
@@ -29,7 +29,7 @@ int allocate_moving_avg(uint32_t num_wbuf, uint32_t subtotal_sizes, moving_avg_v
 
 int reset_moving_avg(moving_avg_values *reset);
 
-int append_to_moving_avg(moving_avg_values *value, int appended, bool *isAtEnd);
+int append_to_moving_avg(moving_avg_values *value, int16_t appended, bool *isAtEnd);
 
 int get_latest_frame_moving_avg(moving_avg_values *value, int *frame);
